requirements.txt
adamp==0.3.0
argcomplete==1.12.3
argon2-cffi==20.1.0
attrs==21.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==4.0.0
brotlipy==0.7.0
cached-property==1.5.2
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1613246939562/work
chardet @ file:///tmp/build/80754af9/chardet_1605303159953/work
click==8.0.1
conda==4.10.1
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262151086/work
configparser==5.0.2
cryptography @ file:///tmp/build/80754af9/cryptography_1616769182610/work
cycler==0.10.0
dataclasses==0.6
debugpy==1.4.1
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
defusedxml==0.7.1
dnspython==2.1.0
docker-pycreds==0.4.0
einops==0.3.0
entrypoints==0.3
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
gitdb==4.0.7
GitPython==3.1.18
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
h5py==3.3.0
huggingface-hub==0.0.12
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
importlib-metadata==4.6.4
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.3
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
linformer==0.2.1
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib==3.4.3
matplotlib-inline==0.1.2
Medutils @ git+https://github.com/nemodleo/medutils.git@fc65b570278fc1fdabc2a80bba456eaab4462aae
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853974840/work
mkl-service==2.3.0
natsort==7.1.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1620831194891/work
olefile==0.46
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1617386154241/work
pkginfo==1.7.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.17.3
psutil @ file:///tmp/build/80754af9/psutil_1612298016854/work
ptflops==0.6.6
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
python-etcd==0.4.5
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.8.3
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016701961/work
sacremoses==0.0.45
scikit-image==0.18.2
scipy==1.7.1
Send2Trash==1.8.0
sentry-sdk==1.3.1
shortuuid==1.0.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==4.0.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1616183228191/work
subprocess32==3.5.4
terminado==0.11.0
testpath==0.5.0
tifffile==2021.8.8
timm==0.4.12
tokenizers==0.10.3
torch==1.9.0
torchaudio==0.9.0
torchelastic==0.2.0
torchtext==0.10.0
torchvision==0.10.0
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
transformers==4.9.2
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wandb==0.12.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
yacs==0.1.8
zipp==3.5.0
